+++ src/Symfony/Component/Translation/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 6.2
 ---
 
+ * Deprecate `PhpStringTokenParser`
  * Deprecate `PhpExtractor` in favor of `PhpAstExtractor`
  * Add `PhpAstExtractor` (requires [nikic/php-parser](https://github.com/nikic/php-parser) to be installed)